services/queue-manager/app/services/arango_service.py
from loguru import logger
from arango import ArangoClient
from app.config import ARANGO_URL, ARANGO_DB, ARANGO_USER, ARANGO_PASSWORD

class ArangoQueueStore:
    def __init__(self):
        self.client = ArangoClient()
        self.db = None

    def connect(self):
        sys = self.client.db("_system", username=ARANGO_USER, password: "REDACTED")
        if not sys.has_database(ARANGO_DB):
            sys.create_database(ARANGO_DB)
            logger.info(f"📁 Создана база данных: {ARANGO_DB}")
        self.db = self.client.db(ARANGO_DB, username=ARANGO_USER, password: "REDACTED")

    def get_collection(self, name: str):
        if not self.db.has_collection(name):
            self.db.create_collection(name)
            logger.info(f"📦 Создана коллекция: {name}")
        return self.db.collection(name)

    async def save_queue_meta(queue_id: str, command: StartQueueCommand):
    # ... (реализация сохранения в ArangoDB)
    logger.info(f"Сохранение метаданных для очереди: {queue_id}")
    # Здесь должна быть реальная логика сохранения в ArangoDB
    # Например, arango_store.insert_document('queues', {'_key': queue_id, **command.dict()})
    pass
