services/queue-manager/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
loguru==0.7.2
python-dotenv==1.0.1
aiohttp==3.9.5
pydantic==2.7.1
aiokafka==0.10.0
python-arango==7.6.1
prometheus-client==0.20.0
-e ../../libs/streamforge_utils